// src/app/api/generated/model/cobertura-para-oferta-practica.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CoberturaParaOfertaArancel } from './cobertura-para-oferta-arancel';


/**
 * Como resolvio una practica de la oferta bajo la cobertura
 */
export interface CoberturaParaOfertaPractica { 
    /**
     * Null si no resolvio
     */
    arancel?: CoberturaParaOfertaArancel;
    /**
     * Por que no resolvio. Null si resolvio
     */
    motivo?: CoberturaParaOfertaPracticaMotivoEnum;
    practicaId?: number;
    /**
     * Si es la practica principal de la oferta
     */
    principal?: boolean;
}
export enum CoberturaParaOfertaPracticaMotivoEnum {
    SIN_CONVENIO_VIGENTE = 'SIN_CONVENIO_VIGENTE',
    SIN_ARANCEL_VIGENTE = 'SIN_ARANCEL_VIGENTE'
};